Get All Grab Sample Data (1)

get Senate's grab sample data (type 1) from all csv files in a directory and return it in a "all-in-one-table" data frame

hsGetAllGsData1(csvdir, sep = ";", dateFormat = "%d.%m.%Y",
  dec = ",", data.block.number = NA, csvFiles = dir(csvdir, "*.csv",
  full.names = TRUE))

Arguments

csvdir
absolute path to directory in which csv files to be read are located.

sep
separator in csv file, e.g. ";" or ","

dateFormat
date format specifier describing the format in which dates are represented in the csv file. Use placeholders , "%d" (day), "%m" (month), "%y" (2-digit year), "%Y" (4-digit year) to describe the date format. "%d.%m.%Y", "%d/%m/%y", "%Y-%m-%d" are examples for valid format specifiers.

dec
decimal character: "." or ","

data.block.number
passed to hsGetGsData1, for a description see there.

csvFiles
vector of full paths to csv files to be read. All files must contain data for the same monitoring points, in the same order! Default: all csv files in csvdir

See also