Read BWB Rain Data

Read BWB Rain Data

readBwbRainData(
  file,
  tableName = "Niederschlag$",
  columnNames = NULL,
  toUTC = TRUE,
  toNormaltime = FALSE,
  dbg = TRUE,
  use2007Driver = kwb.db::isExcel2007File(file),
  ...
)

Arguments

file

full path to Excel file

tableName

sheet name, default: "[Niederschlag$]"

columnNames

optional. Vector of names of columns (after renaming with gaugeNamesShort) to return

toUTC

default: TRUE

toNormaltime

if TRUE, a column tBeg_WT is created containing the begin of the time interval in Normal time (winter time, UTC+01). default: FALSE

dbg

default: TRUE

use2007Driver

if TRUE, 2007 driver is used even if extension of file is ".xls"

...

additional parameters passed to readBwbRainFromCsv