Download and Read Data from wasserportal.berlin.de

This function downloads and reads CSV files from wasserportal.berlin.de.

read_wasserportal(
  station,
  variables = get_wasserportal_variables(station),
  from_date = as.character(Sys.Date() - 90L),
  type = "single",
  include_raw_time = FALSE
)

Arguments

station
station number, as returned by get_wasserportal_stations

variables
vector of variable identifiers, as returned by get_wasserportal_variables

from_date
Date object (or string in format "yyyy-mm-dd" that can be converted to a Date object representing the first day for which to request data. Default: as.character(Sys.Date() - 90L)

type
one of "single" (the default), "daily", "monthly"

include_raw_time
if TRUE the original time column and the column with the corrected winter time are included in the output. The default is FALSE.

Value

data frame read from the CSV file that the download provides. IMPORTANT: It is not yet clear how to interpret the timestamp, see example

Details

The original timestamps (column timestamps_raw in the example below) are not all plausible, e.g. "31.03.2019 03:00" appears twice! They are corrected (column timestamp_corr) to represent a plausible sequence of timestamps in Berlin Normal Time (UTC+01) Finally, a valid POSIXct timestamp in timezone "Berlin/Europe" (UTC+01 in winter, UTC+02 in summer) is created, together with the additional information on the UTC offset (column UTCOffset, 1 in winter, 2 in summer).

Examples

# Get a list of available water quality stations and variables
stations <- kwb.read::get_wasserportal_stations()
variables <- kwb.read::get_wasserportal_variables()

# Read the timeseries (multiple variables for one station)
water_quality <- kwb.read::read_wasserportal(
  station = stations$MPS_Charlottenburg,
  variables = c(variables["Sauerstoffgehalt"], variables["Leitfaehigkeit"]),
  from_date = as.character(Sys.Date() - 30L), include_raw_time = TRUE
)

#> Reading 'Sauerstoffgehalt' for station 153 (MPS_Charlottenburg) ... ok. (0.90s) 
#> Reading 'Leitfaehigkeit' for station 153 (MPS_Charlottenburg) ... ok. (1.46s) 

# Look at the first few records
head(water_quality)

#>         timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 2789 31.01.2021 00:00 31.01.2021 00:00 2021-01-31 00:00:00         1
#> 2790 31.01.2021 00:15 31.01.2021 00:15 2021-01-31 00:15:00         1
#> 2791 31.01.2021 00:30 31.01.2021 00:30 2021-01-31 00:30:00         1
#> 2792 31.01.2021 00:45 31.01.2021 00:45 2021-01-31 00:45:00         1
#> 2793 31.01.2021 01:00 31.01.2021 01:00 2021-01-31 01:00:00         1
#> 2794 31.01.2021 01:15 31.01.2021 01:15 2021-01-31 01:15:00         1
#>      Sauerstoffgehalt Leitfaehigkeit
#> 2789             -777           -777
#> 2790             -777           -777
#> 2791             -777           -777
#> 2792             -777           -777
#> 2793             -777           -777
#> 2794             -777           -777

# Check the metadata
kwb.utils::getAttribute(water_quality, "metadata")

#> $Sauerstoffgehalt
#> [1] "Stationsnummer: 153"              "Stationsname: MPS Charlottenburg"
#> [3] "Gewässer: Spree"                  "Sauerstoffgehalt im mg/l"        
#> [5] "Fehlwerte: -777"                 
#> 
#> $Leitfaehigkeit
#> [1] "Stationsnummer: 153"              "Stationsname: MPS Charlottenburg"
#> [3] "Gewässer: Spree"                  "Leitfähigkeit in µS/cm"          
#> [5] "Fehlwerte: -777"                 
#> 

# Set missing values to NA
water_quality[water_quality == -777] <- NA

# Look at the first few records again
head(water_quality)

#>         timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 2789 31.01.2021 00:00 31.01.2021 00:00 2021-01-31 00:00:00         1
#> 2790 31.01.2021 00:15 31.01.2021 00:15 2021-01-31 00:15:00         1
#> 2791 31.01.2021 00:30 31.01.2021 00:30 2021-01-31 00:30:00         1
#> 2792 31.01.2021 00:45 31.01.2021 00:45 2021-01-31 00:45:00         1
#> 2793 31.01.2021 01:00 31.01.2021 01:00 2021-01-31 01:00:00         1
#> 2794 31.01.2021 01:15 31.01.2021 01:15 2021-01-31 01:15:00         1
#>      Sauerstoffgehalt Leitfaehigkeit
#> 2789               NA             NA
#> 2790               NA             NA
#> 2791               NA             NA
#> 2792               NA             NA
#> 2793               NA             NA
#> 2794               NA             NA

### How was the original timestamp interpreted?

# Determine the days at which summer time starts and ends, respectively
switches <- kwb.datetime::date_range_CEST(2019)

# Reformat to dd.mm.yyyy
switches <- kwb.datetime::reformatTimestamp(switches, "%Y-%m-%d", "%d.%m.%Y")

# Define a pattern to look for timestamps "around" the switches
pattern <- paste(switches, "0[1-4]", collapse = "|")

# Look at the data for these timestamps
water_quality[grepl(pattern, water_quality$timestamp_raw), ]

#> [1] timestamp_raw    timestamp_corr   LocalDateTime    UTCOffset       
#> [5] Sauerstoffgehalt Leitfaehigkeit  
#> <0 rows> (or 0-length row.names)

# The original timestamps (timestamps_raw) are not all plausible, e.g. 
# "31.03.2019 03:00" appears twice! See the Details in ?read_wasserportal()
# how this is treated.