Download and Read Data from wasserportal.berlin.de

Download and Read Data from wasserportal.berlin.de

read_wasserportal_raw(station, from_date = "2019-01-01")

Arguments

station
station number, one of

601

MPS Berlin-Spandauer-Schifffahrtskanal

151

MPS Caprivibruecke

153

MPS Charlottenburg

509

MPS Landwehrkanal

504

MPS Neukoellner Schifffahrtskanal

414

MPS Teltowkanal

141

MS Muehlendammschleuse

111

MS Rahnsdorf

211

MS Schmoeckwitz

161

MS Sophienwerder

421

MS Teltow-Werft

from_date
Date object (or string in format "yyyy-mm-dd" that can be converted to a Date object representing the first day for which to request data

Value

data frame read from the CSV file that the download provides. IMPORTANT: It is not yet clear how to interpret the timestamp, see example

Examples

# Get a list of available water quality stations
stations <- kwb.read::get_wasserportal_stations()
temperature_raw <- kwb.read::read_wasserportal_raw(stations$MPS_Charlottenburg)

# Look at the first few records
head(temperature_raw)

#>              Datum Einzelwert
#> 1 01.01.2019 00:00       -777
#> 2 01.01.2019 00:15       -777
#> 3 01.01.2019 00:30       -777
#> 4 01.01.2019 00:45       -777
#> 5 01.01.2019 01:00       -777
#> 6 01.01.2019 01:15       -777

# Check the metadata
kwb.utils::getAttribute(temperature_raw, "metadata")

#> [1] "Stationsnummer: 153"              "Stationsname: MPS Charlottenburg"
#> [3] "Gewässer: Spree"                  "Wasssertemperatur in °C"         
#> [5] "Fehlwerte: -777"                 

# Set missing values to NA
temperature_raw$Einzelwert[temperature_raw$Einzelwert == -777] <- NA

# Look at the first few records
head(temperature_raw)

#>              Datum Einzelwert
#> 1 01.01.2019 00:00         NA
#> 2 01.01.2019 00:15         NA
#> 3 01.01.2019 00:30         NA
#> 4 01.01.2019 00:45         NA
#> 5 01.01.2019 01:00         NA
#> 6 01.01.2019 01:15         NA

# How to interpret the timestamp?
# Determine the days at which summer time starts and ends, respectivel
switches <- kwb.datetime::date_range_CEST(2019)

# Reformat to dd.mm.yyyy
switches <- kwb.datetime::reformatTimestamp(switches, "%Y-%m-%d", "%d.%m.%Y")

# Define a pattern to look for timestamps "around" the switches
pattern <- paste(switches, "0[1-4]", collapse = "|")

# Look at the data for these timestamps
temperature_raw[grepl(pattern, temperature_raw$Datum), ]

#>                 Datum Einzelwert
#> 8549 31.03.2019 01:00         NA
#> 8550 31.03.2019 01:15         NA
#> 8551 31.03.2019 01:30         NA
#> 8552 31.03.2019 01:45         NA
#> 8553 31.03.2019 03:00         NA
#> 8554 31.03.2019 03:15         NA
#> 8555 31.03.2019 03:30         NA
#> 8556 31.03.2019 03:45         NA
#> 8557 31.03.2019 03:00         NA
#> 8558 31.03.2019 03:15         NA
#> 8559 31.03.2019 03:30         NA
#> 8560 31.03.2019 03:45         NA
#> 8561 31.03.2019 04:00         NA
#> 8562 31.03.2019 04:15         NA
#> 8563 31.03.2019 04:30         NA
#> 8564 31.03.2019 04:45         NA

# The timestamps are not plausible, e.g. "31.03.2019 03:00" appears twice!