Read Rain Data from BWB Excel Files

Read Rain Data from BWB Excel Files

Usage

read_rain(
  directory = kwb.utils::resolve("RAIN_DIR", dictionary),
  pattern = kwb.utils::resolve("RAIN_XLS_PATTERN", dictionary),
  tableName = kwb.utils::resolve("RAIN_XLS_TABLE", dictionary),
  dictionary = kwb.utils::selectElements(settings, "dictionary"),
  rain.xls.files = dir(directory, pattern, full.names = TRUE),
  aggregation.interval = kwb.utils::selectElements(settings, "rain.aggregation.interval"),
  settings = NULL,
  use2007Driver = TRUE
)

Arguments

directory

directory containing excel files from which rain data are read. Default: kwb.utils::resolve("RAIN_DIR", dictionary)

pattern

file name pattern matching names of files to be read. Default: kwb.utils::resolve("RAIN_XLS_PATTERN", dictionary)

tableName

name of sheet in the excel files containing the rain data. Default: kwb.utils::resolve("RAIN_XLS_TABLE", dictionary)

dictionary

list of key/value pairs that can be used as a dictionary to look up the following keywords (if directory, pattern, tableName are not given): RAIN_DIR, RAIN_XLS_PATTERN, RAIN_XLS_TABLE

rain.xls.files

full paths to xls or xlsx files to be read. If given, these paths are used. Otherwise all files in directory matching the filename pattern are used.

aggregation.interval

aggregation interval in seconds. Default: settings$rain.aggregation.interval

settings

list that may contain arguments to this function in elements rain.aggregation.interval or dictionary

use2007Driver

passed to getAggregatedBwbRainDataFromExcelFiles

Value

data frame with column DateTime containing timestamps in UTC timezone but representing Berlin normal (= winter) time.