Download and Read Data from wasserportal.berlin.de

This function downloads and reads CSV files from wasserportal.berlin.de.

read_wasserportal(station,
  variables = get_wasserportal_variables(station),
  from_date = "2019-01-01", include_raw_time = FALSE)

Arguments

station
station number, as returned by get_wasserportal_stations

variables
vector of variable identifiers, as returned by get_wasserportal_variables

from_date
Date object (or string in format "yyyy-mm-dd" that can be converted to a Date object representing the first day for which to request data

include_raw_time
if TRUE the original time column and the column with the corrected winter time are included in the output. The default is FALSE.

Value

data frame read from the CSV file that the download provides. IMPORTANT: It is not yet clear how to interpret the timestamp, see example

Details

The original timestamps (column timestamps_raw in the example below) are not all plausible, e.g. "31.03.2019 03:00" appears twice! They are corrected (column timestamp_corr) to represent a plausible sequence of timestamps in Berlin Normal Time (UTC+01) Finally, a valid POSIXct timestamp in timezone "Berlin/Europe" (UTC+01 in winter, UTC+02 in summer) is created, together with the additional information on the UTC offset (column UTCOffset, 1 in winter, 2 in summer).

Examples

# Get a list of available water quality stations and variables
stations <- kwb.read::get_wasserportal_stations()
variables <- kwb.read::get_wasserportal_variables()

# Read the timeseries (multiple variables for one station)
water_quality <- kwb.read::read_wasserportal(
  station = stations$MPS_Charlottenburg,
  variables = c(variables["Sauerstoffgehalt"], variables["Leitfaehigkeit"]),
  from_date = "2019-03-01", include_raw_time = TRUE
)

#> Reading 'Sauerstoffgehalt' for station 153 (MPS_Charlottenburg) ... ok. (3.00s) 
#> Reading 'Leitfaehigkeit' for station 153 (MPS_Charlottenburg) ... ok. (1.58s) 

# Look at the first few records
head(water_quality)

#>      timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 1 01.03.2019 00:00 01.03.2019 00:00 2019-03-01 00:00:00         1
#> 2 01.03.2019 00:15 01.03.2019 00:15 2019-03-01 00:15:00         1
#> 3 01.03.2019 00:30 01.03.2019 00:30 2019-03-01 00:30:00         1
#> 4 01.03.2019 00:45 01.03.2019 00:45 2019-03-01 00:45:00         1
#> 5 01.03.2019 01:00 01.03.2019 01:00 2019-03-01 01:00:00         1
#> 6 01.03.2019 01:15 01.03.2019 01:15 2019-03-01 01:15:00         1
#>   Sauerstoffgehalt Leitfaehigkeit
#> 1             -777           -777
#> 2             -777           -777
#> 3             -777           -777
#> 4             -777           -777
#> 5             -777           -777
#> 6             -777           -777

# Check the metadata
kwb.utils::getAttribute(water_quality, "metadata")

#> $Sauerstoffgehalt
#> [1] "Stationsnummer: 153"              "Stationsname: MPS Charlottenburg"
#> [3] "Gewässer: Spree"                  "Sauerstoffgehalt im mg/l"        
#> [5] "Fehlwerte: -777"                 
#> 
#> $Leitfaehigkeit
#> [1] "Stationsnummer: 153"              "Stationsname: MPS Charlottenburg"
#> [3] "Gewässer: Spree"                  "Leitfähigkeit in µS/cm"          
#> [5] "Fehlwerte: -777"                 
#> 

# Set missing values to NA
water_quality[water_quality == -777] <- NA

# Look at the first few records again
head(water_quality)

#>      timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 1 01.03.2019 00:00 01.03.2019 00:00 2019-03-01 00:00:00         1
#> 2 01.03.2019 00:15 01.03.2019 00:15 2019-03-01 00:15:00         1
#> 3 01.03.2019 00:30 01.03.2019 00:30 2019-03-01 00:30:00         1
#> 4 01.03.2019 00:45 01.03.2019 00:45 2019-03-01 00:45:00         1
#> 5 01.03.2019 01:00 01.03.2019 01:00 2019-03-01 01:00:00         1
#> 6 01.03.2019 01:15 01.03.2019 01:15 2019-03-01 01:15:00         1
#>   Sauerstoffgehalt Leitfaehigkeit
#> 1               NA             NA
#> 2               NA             NA
#> 3               NA             NA
#> 4               NA             NA
#> 5               NA             NA
#> 6               NA             NA

### How was the original timestamp interpreted?

# Determine the days at which summer time starts and ends, respectively
switches <- kwb.datetime::date_range_CEST(2019)

# Reformat to dd.mm.yyyy
switches <- kwb.datetime::reformatTimestamp(switches, "%Y-%m-%d", "%d.%m.%Y")

# Define a pattern to look for timestamps "around" the switches
pattern <- paste(switches, "0[1-4]", collapse = "|")

# Look at the data for these timestamps
water_quality[grepl(pattern, water_quality$timestamp_raw), ]

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 21387 31.03.2019 01:00 31.03.2019 01:00 2019-03-31 01:00:00         1
#> 21388 31.03.2019 01:15 31.03.2019 01:15 2019-03-31 01:15:00         1
#> 21389 31.03.2019 01:30 31.03.2019 01:30 2019-03-31 01:30:00         1
#> 21390 31.03.2019 01:45 31.03.2019 01:45 2019-03-31 01:45:00         1
#> 21391 31.03.2019 03:00 31.03.2019 02:00 2019-03-31 03:00:00         2
#> 21393 31.03.2019 03:15 31.03.2019 02:15 2019-03-31 03:15:00         2
#> 21395 31.03.2019 03:30 31.03.2019 02:30 2019-03-31 03:30:00         2
#> 21397 31.03.2019 03:45 31.03.2019 02:45 2019-03-31 03:45:00         2
#> 21392 31.03.2019 03:00 31.03.2019 03:00 2019-03-31 04:00:00         2
#> 21394 31.03.2019 03:15 31.03.2019 03:15 2019-03-31 04:15:00         2
#> 21396 31.03.2019 03:30 31.03.2019 03:30 2019-03-31 04:30:00         2
#> 21398 31.03.2019 03:45 31.03.2019 03:45 2019-03-31 04:45:00         2
#> 21399 31.03.2019 04:00 31.03.2019 04:00 2019-03-31 05:00:00         2
#> 21400 31.03.2019 04:15 31.03.2019 04:15 2019-03-31 05:15:00         2
#> 21401 31.03.2019 04:30 31.03.2019 04:30 2019-03-31 05:30:00         2
#> 21402 31.03.2019 04:45 31.03.2019 04:45 2019-03-31 05:45:00         2
#>       Sauerstoffgehalt Leitfaehigkeit
#> 21387               NA             NA
#> 21388               NA             NA
#> 21389               NA             NA
#> 21390               NA             NA
#> 21391               NA             NA
#> 21393               NA             NA
#> 21395               NA             NA
#> 21397               NA             NA
#> 21392               NA             NA
#> 21394               NA             NA
#> 21396               NA             NA
#> 21398               NA             NA
#> 21399               NA             NA
#> 21400               NA             NA
#> 21401               NA             NA
#> 21402               NA             NA

# The original timestamps (timestamps_raw) are not all plausible, e.g. 
# "31.03.2019 03:00" appears twice! See the Details in ?read_wasserportal()
# how this is treated.